refactor code and basic implementation of Context API
