+++ src/context/product-context.js
@@ -1,0 +1,7 @@
+import {createContext} from 'react';
+
+const ProductContext = createContext({
+    count: 0,
+})
+
+export default ProductContext;
